refactor : 테스트 코드 구조 수정
중복되는 테스트 코드 제거, 패키지 이동
