+++ src/test/java/roomescape/controller/ReservationTimeControllerTest.java
@@ -1,7 +1,6 @@
-package roomescape.repository;
+package roomescape.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.DynamicTest.dynamicTest;
 
 import io.restassured.RestAssured;
@@ -20,11 +19,12 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.domain.Reservation;
+
+import roomescape.domain.ReservationTime;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class ReservationRepositoryTest {
+public class ReservationTimeControllerTest {
 
     @LocalServerPort
     int serverPort;
@@ -38,55 +38,52 @@ public class ReservationRepositoryTest {
     private JdbcTemplate jdbcTemplate;
 
     @Test
-    @DisplayName("DB에서 예약 내역 조회 API 작동을 확인한다")
+    @DisplayName("DB에서 예약 시간 목록 조회 API 작동을 확인한다")
     void checkAllReservations() {
-        jdbcTemplate.update("INSERT INTO reservation (name, date, time) VALUES (?, ?, ?)", "브라운", "2023-08-05", "15:40");
+        jdbcTemplate.update("INSERT INTO reservation_time (start_at) VALUES (?)", "15:40");
 
-        List<Reservation> reservations = RestAssured.given().log().all()
-                .when().get("/reservations")
+        List<ReservationTime> reservationTimes = RestAssured.given().log().all()
+                .when().get("/times")
                 .then().log().all()
                 .statusCode(200).extract()
-                .jsonPath().getList(".", Reservation.class);
+                .jsonPath().getList(".", ReservationTime.class);
 
-        Integer count = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
+        Integer count = jdbcTemplate.queryForObject("SELECT count(*) from reservation_time", Integer.class);
 
-        assertThat(reservations.size()).isEqualTo(count);
+        assertThat(reservationTimes.size()).isEqualTo(count);
     }
 
     @TestFactory
-    @DisplayName("DB에서 예약 추가와 삭제의 작동을 확인한다")
+    @DisplayName("DB에서 예약 시간 추가와 삭제의 작동을 확인한다")
     Stream<DynamicTest> reservationCreateAndDelete() {
         Map<String, String> params = Map.of(
-                "name", "브라운",
-                "date", "2023-08-05",
-                "time", "10:00"
+                "startAt", "10:00"
         );
 
         return Stream.of(
-                dynamicTest("예약을 추가한다", () -> {
+                dynamicTest("예약 시간을 추가한다", () -> {
                     RestAssured.given().log().all()
                             .contentType(ContentType.JSON)
                             .body(params)
-                            .when().post("/reservations")
+                            .when().post("/times")
                             .then().log().all()
-                            .statusCode(201)
-                            .header("Location", "/reservations/1");
+                            .statusCode(201);
                 }),
 
-                dynamicTest("예약이 정상적으로 추가되었는지 확인한다", () -> {
-                    Integer count = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
+                dynamicTest("예약 시간이 정상적으로 추가되었는지 확인한다", () -> {
+                    Integer count = jdbcTemplate.queryForObject("SELECT count(*) from reservation_time", Integer.class);
                     assertThat(count).isEqualTo(1);
                 }),
 
                 dynamicTest("id가 1인 예약을 삭제한다", () -> {
                     RestAssured.given().log().all()
-                            .when().delete("/reservations/1")
+                            .when().delete("/times/1")
                             .then().log().all()
                             .statusCode(204);
                 }),
 
                 dynamicTest("예약이 정상적으로 삭제되었는지 확인한다", () -> {
-                    Integer countAfterDelete = jdbcTemplate.queryForObject("SELECT count(1) from reservation", Integer.class);
+                    Integer countAfterDelete = jdbcTemplate.queryForObject("SELECT count(*) from reservation_time", Integer.class);
                     assertThat(countAfterDelete).isEqualTo(0);
                 })
         );
